Strip chart resolution was not working correctly. Fixed.
git-svn-id: svn+ssh://svn.osdn.jp/svnroot/alchemusica/trunk@202 19e2383c-df9d-4eb5-8775-ae43218f5bcb

+++ Classes/StripChartView.h
@@ -41,9 +41,9 @@ enum {
 
     //  Used for showing cursor info when stripDraggingMode > 0
     //  (i.e. while dragging the selected event(s))
-    int initialDraggedValue;    //  The initial strip value of the dragged event
+    float initialDraggedValue;    //  The initial strip value of the dragged event
     int32_t initialDraggedTick; //  The initial tick value of the dragged event
-    int deltaDraggedValue;
+    float deltaDraggedValue;
     int32_t deltaDraggedTick;
     
     //  Resolution of the y value. Usually 1.0, but can be other values.
